requirements.txt
flask==2.3.2
pymongo==4.4.0
python-dotenv==0.21.0
flask-jwt-extended==4.5.2
flask-cors==4.0.0
bcrypt==4.1.2